requirements.txt
alembic==1.4.3
amqp==5.0.1
aniso8601==8.0.0
billiard==3.6.3.0
celery==5.0.0
cffi==1.14.3
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==2.9.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.0.0
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
prompt-toolkit==3.0.7
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.19
vine==5.0.0
waitress==1.4.4
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.2.0
